third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/breakpoint-manager-listeners-count-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that scripts panel does not create too many source frames.


Running: testSourceFramesCount
Page reloaded.
Number of breakpoint-added event listeners is 3
Dumping SourceFrames listening for breakpoint-added event:
    script1.js
    script2.js
